fix(IN-291): 리뷰 반영
src/shared/api/sidebarStore.ts: 글로벌 사이드바 상태임을 명시하는 주석을 추가했습니다. 마이페이지 전용 state로 오해할 수 있어 명확히 했습니다.

src/features/me/ui/MyPageSidebar.tsx: currentPath를 외부에서 prop으로 받던 방식을 제거하고, 컴포넌트 내부에서 usePathname()으로 현재 경로를 직접 조회하도록 변경했습니다.

src/pages/me/ui/MyPageLayout.tsx: 각 마이페이지 컴포넌트에 중복되던 wrapper div, MyPageSidebar, setOpen(false) useEffect를 하나의 공통 layout 컴포넌트로 통합했습니다.
