+++ app/(protected)/me/layout.ts
@@ -1,0 +1,1 @@
+export { MyPageLayout as default } from '@/pages/me'
