+++ src/pages/me/alarm/ui/MyAlarmPage.tsx
@@ -1,20 +1,5 @@
 'use client'
 
-import { useEffect } from 'react'
-import { MyPageSidebar } from '@/features/me'
-import { useSidebarStore } from '@/shared/api'
-
 export function MyAlarmPage() {
-  const setOpen = useSidebarStore((state) => state.setOpen)
-
-  useEffect(() => {
-    setOpen(false)
-  }, [setOpen])
-
-  return (
-    <div className='flex size-full bg-background-gray-default'>
-      <MyPageSidebar currentPath='/me/alarm' />
-      <div className='flex-1 bg-gray-200 px-10 py-8' />
-    </div>
-  )
+  return <div className='flex-1 bg-gray-200 px-10 py-8' />
 }
