+++ src/pages/me/index.ts
@@ -1,0 +1,1 @@
+export { MyPageLayout } from './ui/MyPageLayout'
