+++ src/pages/me/profile/ui/MyProfilePage.tsx
@@ -1,28 +1,16 @@
 'use client'
 
-import { useEffect } from 'react'
-import { MyPageSidebar } from '@/features/me'
-import { useSidebarStore } from '@/shared/api'
 import { PersonalInfoSection } from '@/widgets/me/personalInfo'
 import { CustomInfoSection } from '@/widgets/me/customInfo'
 
 export function MyProfilePage() {
-  const setOpen = useSidebarStore((state) => state.setOpen)
-
-  useEffect(() => {
-    setOpen(false)
-  }, [setOpen])
-
   return (
-    <div className='flex size-full bg-background-gray-default'>
-      <MyPageSidebar currentPath='/me/profile' />
-      <div className='flex h-fit w-full flex-col gap-32 p-24'>
-        <h1 className='text-ibm-title-lg-normal text-text-and-icon-default'>
-          프로필 설정
-        </h1>
-        <PersonalInfoSection />
-        <CustomInfoSection />
-      </div>
+    <div className='flex h-fit w-full flex-col gap-32 p-24'>
+      <h1 className='text-ibm-title-lg-normal text-text-and-icon-default'>
+        프로필 설정
+      </h1>
+      <PersonalInfoSection />
+      <CustomInfoSection />
     </div>
   )
 }
